bookies/requirements.txt
asgiref==3.8.1
Django~=5.0.6
cryptography==44.0.0
Django==5.1.3
django-allauth==65.2.0
django-cors-headers==4.6.0
djangorestframework==3.15.2
sqlparse==0.5.2
tzdata==2024.2
requests==2.31.0
django-allauth~=65.2.0
credentials~=1.1
djangorestframework-simplejwt==5.3.1
PyJWT==2.10.1
requests==2.32.3
rest-framework-simplejwt==0.0.2

